Module sixtyfps::docs

This is a pseudo module which only exist for documentation purposes as a way to show the SixtyFPS documentation as part of rustdoc.

The generated_code module contains an commented example of what is generated from the .60 file

The langref module is the reference documentation for the .60 language.

The widgets and builtin_elements modules contains the documentation of elements usable within the .60 files

The layouting module contains the documentation to position the elements.

Modules

builtin_elements

Builtin Elements

debugging_techniques

Debugging Techniques

generated_code

This module exists only to explain the API of the code generated from .60 design markup. Its described structure is not really contained in the compiled crate.

langref

The .60 language reference

layouting

Positioning and Layout of Elements

widgets

Widgets