Struct sixtyfps::docs::generated_code::SampleComponent

pub struct SampleComponent {}

This an example of the API that is generated for a component in .60 design markup. This may help you understand what functions you can call and how you can pass data in and out. This is the source code:

SampleComponent := Window {
    property<int> counter;
    property<string> user_name;
    callback hello;
    /// ... maybe more elements here
}

Implementations

impl SampleComponent

pub fn new() -> Self

Creates a new instance that is reference counted and pinned in memory.

pub fn get_counter(&self) -> i32

A getter is generated for each property declared at the root of the component. In this case, this is the getter that returns the value of the counter property declared in the .60 design markup.

pub fn set_counter(&self, value: i32)

A setter is generated for each property declared at the root of the component, In this case, this is the setter that sets the value of the counter property declared in the .60 design markup.

pub fn get_user_name(&self) -> SharedString

Returns the value of the user_name property declared in the .60 design markup.

pub fn set_user_name(&self, value: SharedString)

Assigns a new value to the user_name property.

pub fn invoke_hello(&self)

For each callback declared at the root of the component, a function to call that callback is generated. This is the function that calls the hello callback declared in the .60 design markup.

pub fn on_hello(&self, f: impl Fn() + 'static)

For each callback declared at the root of the component, a function connect to that callback is generated. This is the function that registers the function f as callback when the callback hello is emitted. In order to access the component in the callback, you’d typically capture a weak reference obtained using ComponentHandle::as_weak and then upgrade it to a strong reference when the callback is run:

    let sample = SampleComponent::new();
    let sample_weak = sample.as_weak();
    sample.as_ref().on_hello(move || {
        let sample = sample_weak.unwrap();
        sample.as_ref().set_counter(42);
    });