Module sixtyfps::testing

This module contains functions useful for unit tests

Traits

HasWindow

This trait gives access to the underlying Window of a component for the purposes of testing.

Functions

mock_elapsed_time

SixtyFPS animations do not use real time, but use a mocked time. Normally, the event loop update the time of the animation using real time, but in tests, it is more convenient to use the fake time. This function will add some milliseconds to the fake time

send_keyboard_string_sequence

Simulate entering a sequence of ascii characters key by key.

send_mouse_click

Simulate a mouse click

set_current_keyboard_modifiers

Simulate a change in keyboard modifiers being pressed

set_window_scale_factor

Applies the specified scale factor to the window that’s associated with the given component. This overrides the value provided by the windowing system.