Trait sixtyfps::ComponentHandle

pub trait ComponentHandle {
    fn as_weak(&self) -> Weak<Self>
    where
        Self: Sized;

    fn clone_strong(&self) -> Self;

    fn show(&self);

    fn hide(&self);

    fn run(&self);
}

This trait describes the common public API of a strongly referenced SixtyFPS component. It allows creating strongly-referenced clones, a conversion into/ a weak pointer as well as other convenience functions.

This trait is implemented by the generated component

Required methods

fn as_weak(&self) -> Weak<Self> where
Self: Sized,

Returns a new weak pointer.

fn clone_strong(&self) -> Self

Returns a clone of this handle that’s a strong reference.

fn show(&self)

Marks the window of this component to be shown on the screen. This registers the window with the windowing system. In order to react to events from the windowing system, such as draw requests or mouse/touch input, it is still necessary to spin the event loop, using crate::run_event_loop.

fn hide(&self)

Marks the window of this component to be hidden on the screen. This de-registers the window from the windowing system and it will not receive any further events.

fn run(&self)

This is a convenience function that first calls Self::show, followed by crate::run_event_loop() and Self::hide.